Remove `\X` characters in character ranges for url regex.
This might not be the right course of action, but removing them gets all
tests to pass. If this is wrong, it implies that the tests are not yet
comprehensive enough, but then again, regex tests almost can't be.

In the original regex set linked to from the documentation
(https://gist.github.com/dperini/729294), these characters are unicode
characters like `\u00a1`. I haven't been able to find the equivalent in
my searches for regex characters in elixir beginning with a `\X`. I'm
sure it also makes for poor search queries.

Closes #9