Add relatively basic but complete documentation for Validator.
Examples are given in the documentation, but doctests have been disabled
because of the complex setup required for a valid struct and changeset
to use. Instead, we just magically assume one is there for us.

Also, set the index page of the documentation to the README. Personally,
I find this helpful as a springboard into the rest of the documentation,
but feel free to change it.

Note that the explanation around the 3rd argument to our fuctions
(`opts`) is kind of weak. I'm not sure what options are supported and
what they do based on what's here. As far as I can tell, it allows you
to customize the invalid message, so I just left it at that. Feel free
to bolster it.

Closes #8